requirements.txt
gunicorn
Django
arrow
asgiref
certifi
chardet
click
django-cors-headers
django-extensions
djangorestframework
djangorestframework-simplejwt
idna
numerapi
numpy
pip
python-dateutil
pytz
pyzmq
requests
setuptools
six
sqlparse
tqdm
urllib3
wheel